send sensor data to iot hub
