+++ src/secrets.h.example
@@ -1,6 +1,6 @@
 #pragma once
 
-// change the file extension by removing ".example" and fill in correct values below:
+// copy this file and change the file extension by removing ".example" and fill in correct values below:
 
 // Wifi
 #define IOT_CONFIG_WIFI_SSID "SSID"
@@ -14,19 +14,3 @@
 
 // use this command to monitor telemetry:
 // az iot hub monitor-events --login "IOTHUB_CONNECTION_STRING" --device-id IOT_CONFIG_DEVICE_ID
-
-// When developing for your own Arduino-based platform,
-// please follow the format '(ard;<platform>)'.
-#define IOT_CONFIG_AZURE_SDK_CLIENT_USER_AGENT "c/" AZ_SDK_VERSION_STRING "(ard;nanorp2040connect)"
-
-// Publish 1 message every 5 seconds
-#define IOT_CONFIG_TELEMETRY_FREQUENCY_MS 5000
-
-// How long a SAS token generated by the sample will be valid, in minutes.
-// The sample will stop working after the SAS token is expired, requiring the device to be reset.
-#define IOT_CONFIG_SAS_TOKEN_EXPIRY_MINUTES 60
-
-// Time Zone Offset
-#define IOT_CONFIG_TIME_ZONE -8
-#define IOT_CONFIG_TIME_ZONE_DAYLIGHT_SAVINGS_DIFF 1
-#define IOT_CONFIG_DAYLIGHT_SAVINGS true
